# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StoryForge MVP API
  version: 0.3.0
  description: MVP API — avatar upload (JPG/PNG ≤5MB, thumb 200²), Google OAuth (JWT sf_session), admin moderation, chat SSE (heartbeat + timeout), snapshots, stories, branches export annotate, profile, account deletion (§2.3), moderation, export, feed (author_display for deleted users). Production mutations on /api/* Origin-guard (middleware). Admins STORYFORGE_ADMIN_USER_IDS.
servers:
  - url: http://localhost:3000
paths:
...
  /api/stories/{id}:
    get:
      summary: Public story detail (author_display is 已注销用户 when author account is deleted)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "404":
          description: Not found
    patch:
      summary: Update story
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                summary:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Success
...
  /api/stories/{id}/relations:
    get:
      summary: List character relations for story (author only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
    post:
      summary: Add relation between two character cards
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
  /api/stories/{id}/relations/{relationId}:
    patch:
      summary: Update relation type or description
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: relationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
    delete:
      summary: Delete relation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: relationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success